aux: writeerror_dev  checks for delay target

+++ test/lib/aux.sh
@@ -1137,6 +1137,11 @@ writeerror_dev() {
 	local name=${PREFIX}-errordev
 
 	if test ! -e ERR_DEV; then
+		# delay target is used for error mapping
+		if test ! -f HAVE_DM_DELAY ; then
+			target_at_least dm-delay 1 1 0 || return 0
+			touch HAVE_DM_DELAY
+		fi
 		dmsetup create -u "TEST-$name" "$name" --table "0 4611686018427387904 error"
 		# Take major:minor of our error device
 		echo "$name" > ERR_DEV_NAME
